Update to patina-devops v0.20 (#48)
## Description

Updates the repository to patina-devops v0.2.0 which changes the release
process as described in the release notes of patina-devops:

ref:
https://github.com/OpenDevicePartnership/patina-devops/releases/tag/v0.2.0

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

N/A

## Integration Instructions

The new release process is to manually trigger the `Create Crate Version
Update PR` action. Once the created PR is reviewed, approved, and
merged, the release on github and crates.io will happen automatically.